backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0
motor==3.3.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Date/Time
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# RAG Dependencies (for Project-Scoped RAG)
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
tiktoken==0.5.2

# Optional: for better PDF extraction
pdfplumber==0.10.3

# Data Source RAG Dependencies
# sentence-transformers for local CPU embedding (all-MiniLM-L6-v2)
sentence-transformers==2.2.2
numpy>=1.21.0
